Ref: set version as constant
It's a best practice to declare a contract version as a constant since it's
a piece of data that will not be changed. The constant keyword will tell
the compiler to store the value of this variable in the bytecode. This
will save you gas whenever you read/retrieve this value.